requirements.txt
altgraph==0.16.1
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
future==0.18.2
google-api-python-client==1.7.11
google-auth==1.7.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
httplib2==0.14.0
idna==2.8
newsapi-python==0.2.5
oauthlib==3.1.0
pefile==2019.4.18
Pillow==6.2.1
pyasn1==0.4.7
pyasn1-modules==0.2.7
PyInstaller==3.5
pypiwin32==223
PySocks==1.7.1
python-dateutil==2.8.1
pywin32==227
pywin32-ctypes==0.2.0
requests==2.21.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
spotipy==2.4.4
timeago==1.0.10
uritemplate==3.0.0
urllib3==1.24.3
